% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.r
\name{add_colgroup}
\alias{add_colgroup}
\title{Add column group}
\usage{
add_colgroup(object, object2, position = "left")
}
\arguments{
\item{object}{A layout object including a body containing the table.}

\item{object2}{A layout object including a body containing the table. The column used for matching is not added.}

\item{position}{A character value 'left' or 'right' specifying where the new columns should be added. Any specification other than 'right' will be interpreted as 'left'.}

\item{by}{A character element with the name of the column specifying the column to use for joining (must be named identically in the layout object and the data frame containing the columns to be added).}
}
\value{
A layout object with new column group added.
}
\description{
\code{add_colgroup} Adds a layout objet to another layout object, as a new column group.
}
\examples{
add_colgroup(example_layout, example_layout, position = 'left')
}
\seealso{
\code{\link{prepare}} for making a layout object.

Other table layout: \code{\link{add_cols}},
  \code{\link{add_rowgroup}}, \code{\link{add_units}},
  \code{\link{firstrow_css}}, \code{\link{layout_html}},
  \code{\link{order_by_vector}},
  \code{\link{order_layout}}, \code{\link{prepare}},
  \code{\link{print_abbrev}}, \code{\link{query_abbrev}}
}
\concept{table layout}
