% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.r
\name{add_cols}
\alias{add_cols}
\title{Add columns}
\usage{
add_cols(object, DF, by, position = "left")
}
\arguments{
\item{object}{A layout object including a body containing the table.}

\item{DF}{A data frame with the columns to be added. The column used for matching is not added.}

\item{by}{A character element with the name of the column specifying the column to use for joining (must be named identically in the layout object and the data frame containing the columns to be added).}

\item{position}{A character value 'left' or 'right' specifying where the new columns should be added. Any specification other than 'right' will be interpreted as 'left'.}
}
\value{
A layout object with new columns added to the object body.
}
\description{
\code{add_cols} Adds a data frame to a layout object table body.
}
\examples{
example_layout <- add_cols(
   example_layout,
   dta_valid_n,
   by = "key",
   position = "left")
}
\seealso{
\code{\link{prepare}} for making a layout object.

Other table layout: \code{\link{add_colgroup}},
  \code{\link{add_rowgroup}}, \code{\link{firstrow_css}},
  \code{\link{layout_html}}, \code{\link{order_by_vector}},
  \code{\link{order_layout}}, \code{\link{prepare}},
  \code{\link{print_abbrev}}, \code{\link{query.abbrev}}
}
