% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.r
\name{add_rowgroup}
\alias{add_rowgroup}
\title{Add row groups}
\usage{
add_rowgroup(object, object2, rgroup_names = "")
}
\arguments{
\item{object}{A layout object including a body containing the table.}

\item{object2}{A layout object including a body containing in its @body the row group to be added.}

\item{rgroup_names}{A character vector specifying the names (labels) to be used for the row groups.}
}
\value{
An S4 object with rows- and row group specifications added.
}
\description{
\code{add_rowgroup} Adds a row group to a layout object and assigns separate row group headers displayed when printing as html. The original object@body will be used as the first row group.
}
\examples{
example_layout <- add_rowgroup(example_layout,
                              example_layout_width,
                              rgroup_names = c("Sepal Length", "Sepal Width"))
}
\seealso{
\code{\link{prepare}} for making a layout object and \code{\link{layout_html}} for printing a html table from such an object.

Other table layout: \code{\link{add_colgroup}},
  \code{\link{add_cols}}, \code{\link{firstrow_css}},
  \code{\link{layout_html}}, \code{\link{order_by_vector}},
  \code{\link{order_layout}}, \code{\link{prepare}},
  \code{\link{print_abbrev}}, \code{\link{query.abbrev}}
}
