% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_helpers.r
\name{add_units}
\alias{add_units}
\title{Add units to labels}
\usage{
add_units(x, y)
}
\arguments{
\item{x}{contains variable labels.}

\item{y}{contains units to be added, matched by name.}
}
\value{
Named vector with labels and added units, if available (otherwise "NA" is added).
}
\description{
\code{add_units} adds units to variable labels from a named vector matching the names of the labels.
}
\examples{
variable_labels <- c(
   'apples' = "Apples",
   'bnanas' = "Bananas",
   'peas' = "Peas"
)
unit_labels <- c(
   'apples' = "pcs",
   'peas' = "kg"
)
add_units(variable_labels, unit_labels)
}
\seealso{
Other table layout: \code{\link{add_colgroup}},
  \code{\link{add_cols}}, \code{\link{add_rowgroup}},
  \code{\link{firstrow_css}}, \code{\link{layout_html}},
  \code{\link{order_by_vector}},
  \code{\link{order_layout}}, \code{\link{prepare}},
  \code{\link{print_abbrev}}, \code{\link{query_abbrev}}
}
\concept{table layout}
