% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.r
\name{firstrow_css}
\alias{firstrow_css}
\title{Format first row of each rowgroup}
\usage{
firstrow_css(object, css_style = "font-size: 2em;", ...)
}
\arguments{
\item{object}{A layout object.}

\item{css_style}{A character vector with the css style options that arfe to be applied. The default css applied is "font -size: 2em;".}
}
\value{
A matrix specifying the css for each cell in the html table.
}
\description{
\code{firstrow_css} makes a css for first row of each rowgroup in a layout object. The output can be fed to the html function as a css.cell argument.
}
\examples{
firstrow_css(example_layout)
}
\seealso{
\code{\link{prepare}} for making a layout object and \code{\link{layout_html}} for printing a layout object to html.

Other table layout: \code{\link{add_colgroup}},
  \code{\link{add_cols}}, \code{\link{add_rowgroup}},
  \code{\link{add_units}}, \code{\link{layout_html}},
  \code{\link{order_by_vector}},
  \code{\link{order_layout}}, \code{\link{prepare}},
  \code{\link{print_abbrev}}, \code{\link{query_abbrev}}
}
