\name{layout-package}
\alias{layout-package}
\alias{layout}
\docType{package}
\title{
(Multilevel) layout algorithms for graph drawing. 
}
\description{
  Algorithms for fast drawing of undirected graphs using force directed layouts optinally with a multi-scale approach.
}
\details{
\tabular{ll}{
Package: \tab layout\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-23\cr
License: \tab BSD\cr
}
 The package provides a wrapper for the prefuse-drawing algorithm (\url{http://prefuse.org/}) see 
 \link{layout.prefuse}. Further a multi-scale graph drawing approach is also possible  \link{layout.multi.level}
}
\author{
Oliver D\"urr <oliver.duerr@zhaw.ch>
}
\references{
  See: \url{http://arxiv.org/abs/1207.6282} for the multi-scale approach.
}
