% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.r
\name{layout_html}
\alias{layout_html}
\title{Html table}
\usage{
layout_html(object, caption = NULL, header = NULL, footer = NULL,
  abbreviations = NULL, labels = NULL, signif_digits = NULL,
  rgroup = NULL, n.rgroup = NULL, ...)
}
\arguments{
\item{object}{A layout object including a body containing the table.}

\item{caption}{An optional character element for table caption.}

\item{header}{An optional character vector used for column headers. If header is not specified an there are row groups in the table object, the header will be blank (ie only row group headers displayed).}

\item{footer}{An optional character element for table footer. Abbreviations, if used, will also be added to the footer.}

\item{abbreviations}{An optional named vector referenced by the abbreviations query which matches abbreviations in the table body, variables column, header, column groups and row headers and adds the corresponding explanations to the table footer.}

\item{labels}{An optional named vector which will be used for replacing  names in variables column, header, column groups and row headers.}

\item{signif_digits}{An optional numeric element. If specified, all numeric columns in table body will be rounded to the specified number of significant digits.}

\item{rgroup}{An optional character vector specifying replacement row group headers. The default is "Group n".}

\item{rgroup}{An optional numeric vector specifying replacement length for row groups.}
}
\value{
A html table.
}
\description{
\code{layout_html} prints layout object as a html table.
}
\examples{
html_out <- layout_html(example_layout,
    ## rgroup=rgr,
    css.rgroup="align-content: center;font-weight: bold;font-size: 1em;",
    caption='cap',
    footer='footer here',
    abbreviations=abbrevs,
    labels=varnames,
    signif_digits=3
    ##css.cell=subheader_css(test3, "align-content: center;font-size: 2em;")
    )
}
\seealso{
\code{\link{prepare}} for making a layout object.

Other table layout: \code{\link{add_colgroup}},
  \code{\link{add_cols}}, \code{\link{add_rowgroup}},
  \code{\link{firstrow_css}},
  \code{\link{order_by_vector}},
  \code{\link{order_layout}}, \code{\link{prepare}},
  \code{\link{print_abbrev}}, \code{\link{query.abbrev}}
}
