% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_helpers.r
\name{order_by_vector}
\alias{order_by_vector}
\title{Order data frame by vector}
\usage{
order_by_vector(x, match_by, order_by)
}
\arguments{
\item{x}{Data frame to order.}

\item{match_by}{Character element specifying name of column to match by.}

\item{order_by}{Character vector with character elements in the desired order, matching those in the dataframe column that is to be matched on.}
}
\value{
Reordered data frame.
}
\description{
\code{order_by_vector} orders a data frame by a character vector.
}
\examples{
None yet.
}
\seealso{
\code{\link{query_abbrev}} for querying an object for abbrevs

Other table layout: \code{\link{add_colgroup}},
  \code{\link{add_cols}}, \code{\link{add_rowgroup}},
  \code{\link{add_units}}, \code{\link{firstrow_css}},
  \code{\link{layout_html}}, \code{\link{order_layout}},
  \code{\link{prepare}}, \code{\link{print_abbrev}},
  \code{\link{query_abbrev}}
}
