% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_helpers.r
\name{print_abbrev}
\alias{print_abbrev}
\title{Print abbreviations}
\usage{
print_abbrev(x)
}
\arguments{
\item{x}{Object to test.}

\item{y}{Named list of abbrevs.}
}
\value{
Character string with each abbrev and explanation pair separated by semicolon.
}
\description{
\code{print_abbrev} prints abbrevs provided in named list, for table footer.
}
\examples{
None yet.
}
\seealso{
\code{\link{query_abbrev}} for querying an object for abbrevs

Other table layout: 
\code{\link{add_colgroup}()},
\code{\link{add_cols}()},
\code{\link{add_rowgroup}()},
\code{\link{add_units}()},
\code{\link{firstrow_css}()},
\code{\link{layout_html}()},
\code{\link{order_by_vector}()},
\code{\link{order_layout}()},
\code{\link{prepare}()},
\code{\link{query_abbrev}()}
}
\concept{table layout}
