% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_helpers.r
\name{query_abbrev}
\alias{query_abbrev}
\title{Query abbreviations}
\usage{
query_abbrev(x, y)
}
\arguments{
\item{x}{Object to test.}

\item{y}{Named list of abbrevs.}
}
\value{
Named list of abbreviations in object and corresponding explanation
}
\description{
\code{query_abbrev} queries if any of the names in a named list is present in an object including object names. Returns named list with all abbreviations present (via grepl).
}
\examples{
None yet.
}
\seealso{
\code{\link{html.layout}} for html tables and \code{\link{print_abbrev}} for printing abbrevs to table footer.

Other table layout: \code{\link{add_colgroup}},
  \code{\link{add_cols}}, \code{\link{add_rowgroup}},
  \code{\link{firstrow_css}}, \code{\link{layout_html}},
  \code{\link{order_by_vector}},
  \code{\link{order_layout}}, \code{\link{prepare}},
  \code{\link{print_abbrev}}
}
